refactor: Reorganized files in Media folder for better structure
Improve code maintainability and readability.
