+++ app/media/[id]/components/AnimeEpisodesContainer/components/EpisodeBySource/component.module.css
@@ -87,7 +87,7 @@
     height: 100%;
     background: center;
     background-repeat: no-repeat;
-    background-image: url("../../../../../public/assets/play-circle-fill.svg");
+    background-image: url("../../../../../../../public/assets/play-circle-fill.svg");
     background-size: 20%;
 }
