+++ app/media/[id]/components/AnimeEpisodesContainer/components/EpisodeBySource/gogoanime.tsx
@@ -3,13 +3,14 @@ import React from 'react'
 import Image from 'next/image'
 import styles from "./component.module.css"
 import placeholderImg from "@/public/photo-placeholder.jpg"
-import ButtonMarkEpisodeAsWatched from '../../../../components/ButtonMarkEpisodeAsWatched'
-import { EpisodeAnimeWatch } from '@/app/ts/interfaces/apiAnimewatchInterface'
+import { MediaEpisodes } from '@/app/ts/interfaces/apiGogoanimeDataInterface'
+import ButtonMarkEpisodeAsWatched from '@/app/components/ButtonMarkEpisodeAsWatched'
 import { motion } from 'framer-motion'
 
 type ComponentTypes = {
-    data: EpisodeAnimeWatch,
+    data: MediaEpisodes,
     mediaId: number,
+    title: string,
     episodeNumber: number,
     backgroundImg?: string,
     episodeDescription?: string,
@@ -21,16 +22,12 @@ type ComponentTypes = {
     }[]
 }
 
-function AniwatchEpisode({ data, mediaId, backgroundImg, episodeDescription, motionStyle, episodesWatched, episodeNumber }: ComponentTypes) {
+function GoGoAnimeEpisode({ data, mediaId, title, backgroundImg, episodeDescription, motionStyle, episodesWatched, episodeNumber }: ComponentTypes) {
 
     return (
         <motion.li className={styles.container} variants={motionStyle} initial="initial" animate="animate" exit="initial">
 
-            {data.isFiller && (
-                <span className={styles.alert_filler_text}>Filler</span>
-            )}
-
-            <Link href={`/watch/${mediaId}?source=aniwatch&episode=${data.number}&q=${data.episodeId}`} className={styles.img_container}>
+            <Link href={`/watch/${mediaId}?source=gogoanime&episode=${data.number}&q=${data.id}`} className={styles.img_container}>
                 <Image
                     src={backgroundImg || placeholderImg}
                     data-other-source={true}
@@ -43,15 +40,15 @@ function AniwatchEpisode({ data, mediaId, backgroundImg, episodeDescription, mot
             </Link>
 
             <div className={styles.title_button_container}>
-                <h3 title={`Episode ${data.number} ${data.title && `- ${data.title}`}`}>
-                    <Link href={`/watch/${mediaId}?source=aniwatch&episode=${data.number}&q=${data.episodeId}`}>
-                        {`${data.number} ${data.title && `- ${data.title}`}`}
+                <h3 title={title ? `Episode ${data.number} - ${title}` : `Episode ${data.number}`}>
+                    <Link href={`/watch/${mediaId}?source=gogoanime&episode=${data.number}&q=${data.id}`}>
+                        {title ? `${data.number} - ${title}` : `Episode ${data.number}`}
                     </Link>
                 </h3>
 
                 <ButtonMarkEpisodeAsWatched
                     episodeNumber={episodeNumber}
-                    episodeTitle={data.title}
+                    episodeTitle={`${data.number}`}
                     mediaId={mediaId}
                     wasWatched={
                         episodesWatched?.find(
@@ -59,6 +56,7 @@ function AniwatchEpisode({ data, mediaId, backgroundImg, episodeDescription, mot
                         ) ? true : false
                     }
                 />
+
             </div>
 
             {episodeDescription && (
@@ -69,4 +67,4 @@ function AniwatchEpisode({ data, mediaId, backgroundImg, episodeDescription, mot
     )
 }
 
-export default AniwatchEpisode
+export default GoGoAnimeEpisode
